refactor(db): move API token store ownership
